Improve Migration
