+++ db/migrate.go
@@ -7,68 +7,7 @@ import (
 	"gopkg.in/yaml.v2"
 	"io/ioutil"
 	"log"
-	"time"
 )
-
-var Events = []model.Event{
-	model.Event{
-		ID: 1,
-		Name: "イベント名1",
-		ImagePath: "img/150x150.png",
-		Capacity: 250,
-		StartDate: time.Now(),
-		Description: "イベント1の詳細イベント1の詳細イベント1の詳細イベント1の詳細イベント1の詳細イベント1の詳細イベント1の詳細イベント1の詳細イベント1の詳細イベント1の詳細イベント1の詳細イベント1の詳細イベント1の詳細イベント1の詳細イベント1の詳細イベント1の詳細イベント1の詳細イベント1の詳細イベント1の詳細イベント1の詳細イベント1の詳細",
-	},
-	model.Event{
-		ID: 2,
-		Name: "イベント名2",
-		ImagePath: "img/150x150.png",
-		Capacity: 250,
-		StartDate: time.Now(),
-		Description: "イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細",
-	},
-	model.Event{
-		ID: 3,
-		Name: "イベント名2",
-		ImagePath: "img/150x150.png",
-		Capacity: 250,
-		StartDate: time.Now(),
-		Description: "イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細",
-	},
-	model.Event{
-		ID: 4,
-		Name: "イベント名2",
-		ImagePath: "img/150x150.png",
-		Capacity: 250,
-		StartDate: time.Now(),
-		Description: "イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細",
-	},
-	model.Event{
-		ID: 5,
-		Name: "イベント名2",
-		ImagePath: "img/150x150.png",
-		Capacity: 250,
-		StartDate: time.Now(),
-		Description: "イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細",
-	},
-	model.Event{
-		ID: 6,
-		Name: "イベント名2",
-		ImagePath: "img/150x150.png",
-		Capacity: 250,
-		StartDate: time.Now(),
-		Description: "イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細",
-	},
-	model.Event{
-		ID: 7,
-		Name: "イベント名2",
-		ImagePath: "img/150x150.png",
-		Capacity: 250,
-		StartDate: time.Now(),
-		Description: "イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細イベント2の詳細",
-	},
-}
-
 
 func main() {
 	yml, err := ioutil.ReadFile("conf/db.yaml")
@@ -98,11 +37,4 @@ func main() {
 		db.DropTable(&model.Event{})
 	}
 	db.Set("gorm:table_options", "ENGINE=InnoDB DEFAULT CHARSET=utf8").CreateTable(&model.Event{})
-
-	log.Println("SkillCategory INSERT")
-	for key := range Events {
-		db.Create(&Events[key])
-	}
-	
-	
 }
